requirements.txt
# Murmur — Dependencies

# OpenAI API (Whisper + GPT-5.4 Nano)
openai>=1.0.0

# Audio capture
sounddevice>=0.4.6
numpy>=1.24.0

# UI
PyQt6>=6.5.0

# macOS hotkeys & system integration
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
pynput>=1.7.6

# Environment
python-dotenv>=1.0.0
